remove sidebar style
