+++ src/containers/sidebar/sidebar.style.js
@@ -1,0 +1,16 @@
+const style = {
+  closeIcon: {
+    display: 'flex',
+    justifyContent: 'end',
+    margin: '10px 10px 10px 260px'
+  },
+  list: { p: 0 },
+  listItem: {
+    padding: '10px 8px',
+    fontSize: '18px',
+    color: 'primary.900',
+    textDecoration: 'none'
+  }
+}
+
+export default style
